Add PartitionedScan variant for allNodesScan
Doesn't really add anything new just exposes a `PartitionedScan<T>`, instead of a `Scan<T>`.
Behind the scenes we use the same basic mechanism except that the new variant ignores
nodes in the transaction state just as the other partitioned-scan variants.

+++ community/kernel-api/src/main/java/org/neo4j/storageengine/api/StorageEntityScanCursor.java
@@ -37,7 +37,7 @@ public interface StorageEntityScanCursor<S extends Scan> extends StorageEntityCu
      * @param sizeHint the batch will try to read this number of entities.
      * @return <code>true</code> if there are entities to be found, otherwise <code>false</code>
      */
-    boolean scanBatch(S scan, int sizeHint);
+    boolean scanBatch(S scan, long sizeHint);
 
     /**
      * Initializes this cursor so that the next call to {@link #next()} will place this cursor at that entity.
