+++ community/kernel/src/main/java/org/neo4j/kernel/impl/newapi/DefaultNodeCursor.java
@@ -90,7 +90,7 @@ class DefaultNodeCursor extends TraceableCursor<DefaultNodeCursor> implements No
     boolean scanBatch(
             Read read,
             AllNodeScan scan,
-            int sizeHint,
+            long sizeHint,
             LongIterator addedNodes,
             boolean hasChanges,
             AccessMode accessMode) {
